history/requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pika==1.2.0
pymongo==4.0.1
python-dotenv==0.19.2
Werkzeug==2.0.2
